+++ cpp/include/qmc/interacting.hpp
@@ -2,6 +2,7 @@
 #define QMC_INTERACTING_HPP
 
 #include "qmc/continuous_configuration.hpp"
+#include "qmc/continuous_observables.hpp"
 #include "qmc/interacting_model.hpp"
 #include "qmc/interacting_sampler.hpp"
 #include "qmc/interaction.hpp"